wip: splitting main package
