+++ main.go
@@ -2,116 +2,23 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"flag"
-	"fmt"
-	"net"
-	"net/http"
 	_ "net/http/pprof"
+	"os"
+	"os/signal"
 	"sync"
+	"syscall"
 	"time"
 
-	tink "github.com/tinkerbell/tink/protos/hardware"
-
-	grpc_middleware "github.com/grpc-ecosystem/go-grpc-middleware"
-	grpc_prometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
-	cacherClient "github.com/packethost/cacher/client"
-	"github.com/packethost/cacher/protos/cacher"
-	"github.com/packethost/hegel/grpc/hegel"
+	grpcserver "github.com/packethost/hegel/grpc-server"
+	httpserver "github.com/packethost/hegel/http-server"
 	"github.com/packethost/hegel/metrics"
-	"github.com/packethost/hegel/xff"
-	"github.com/packethost/pkg/env"
 	"github.com/packethost/pkg/log"
-	"github.com/pkg/errors"
-	tinkClient "github.com/tinkerbell/tink/client"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
 )
 
-type server struct {
-	log            log.Logger
-	hardwareClient hardwareGetter
-
-	subLock       sync.RWMutex
-	subscriptions map[string]*subscription
-}
-
-type hardwareGetter interface {
-	ByIP(ctx context.Context, in getRequest, opts ...grpc.CallOption) (hardware, error)
-	Watch(ctx context.Context, in getRequest, opts ...grpc.CallOption) (watchClient, error)
-}
-
-type getRequest interface{}
-type hardware interface{}
-
-type subscription struct {
-	ID           string        `json:"id"`
-	IP           string        `json:"ip"`
-	InitDuration time.Duration `json:"init_duration"`
-	StartedAt    time.Time     `json:"started_at"`
-	cancel       func()
-	updateChan   chan []byte
-}
-
-type hardwareGetterCacher struct {
-	client cacher.CacherClient
-}
-
-type hardwareGetterTinkerbell struct {
-	client tink.HardwareServiceClient
-}
-
-func (hg hardwareGetterCacher) ByIP(ctx context.Context, in getRequest, opts ...grpc.CallOption) (hardware, error) {
-	hw, err := hg.client.ByIP(ctx, in.(*cacher.GetRequest), opts...)
-	if err != nil {
-		return nil, err
-	}
-	return hw, nil
-}
-
-func (hg hardwareGetterCacher) Watch(ctx context.Context, in getRequest, opts ...grpc.CallOption) (watchClient, error) {
-	w, err := hg.client.Watch(ctx, in.(*cacher.GetRequest), opts...)
-	if err != nil {
-		return nil, err
-	}
-	return w, nil
-}
-
-func (hg hardwareGetterTinkerbell) ByIP(ctx context.Context, in getRequest, opts ...grpc.CallOption) (hardware, error) {
-	hw, err := hg.client.ByIP(ctx, in.(*tink.GetRequest), opts...)
-	if err != nil {
-		return nil, err
-	}
-	return hw, nil
-}
-
-func (hg hardwareGetterTinkerbell) Watch(ctx context.Context, in getRequest, opts ...grpc.CallOption) (watchClient, error) {
-	w, err := hg.client.Watch(ctx, in.(*tink.GetRequest), opts...)
-	if err != nil {
-		return nil, err
-	}
-	return w, nil
-}
-
 var (
-	facility = flag.String("facility", env.Get("HEGEL_FACILITY", "onprem"),
-		"The facility we are running in (mostly to connect to cacher)")
-	tlsCertPath = flag.String("tls_cert", env.Get("HEGEL_TLS_CERT"),
-		"Path of tls certificat to use.")
-	tlsKeyPath = flag.String("tls_key", env.Get("HEGEL_TLS_KEY"),
-		"Path of tls key to use.")
-	useTLS = flag.Bool("use_tls", env.Bool("HEGEL_USE_TLS", true),
-		"Whether we should use tls or not (should be disabled for traefik)")
-	metricsPort = flag.Int("http_port", env.Int("HEGEL_HTTP_PORT", 50061),
-		"Port to liten on http")
-	customEndpoints     string
-	gitRev              string = "undefind"
-	gitRevJSON          []byte
-	StartTime           = time.Now()
-	logger              log.Logger
-	hegelServer         *server
-	isCacherAvailableMu sync.RWMutex
-	isCacherAvailable   bool
+	gitRev string = "undefind"
+	logger log.Logger
 )
 
 func main() {
@@ -127,130 +34,47 @@ func main() {
 
 	metrics.State.Set(metrics.Initializing)
 
-	port := env.Int("GRPC_PORT", 42115)
-
-	serverOpts := make([]grpc.ServerOption, 0)
-
-	// setup tls credentials
-	if *useTLS {
-		creds, err := credentials.NewServerTLSFromFile(*tlsCertPath, *tlsKeyPath)
-		if err != nil {
-			logger.Error(err, "failed to initialize server credentials")
-			panic(err)
-		}
-		serverOpts = append(serverOpts, grpc.Creds(creds))
-	}
-
-	trustedProxies := xff.ParseTrustedProxies()
-	xffStream, xffUnary := xff.GRPCMiddlewares(logger, trustedProxies)
-	streamLogger, unaryLogger := logger.GRPCLoggers()
-	serverOpts = append(serverOpts,
-		grpc_middleware.WithUnaryServerChain(
-			xffUnary,
-			unaryLogger,
-			grpc_prometheus.UnaryServerInterceptor,
-		),
-		grpc_middleware.WithStreamServerChain(
-			xffStream,
-			streamLogger,
-			grpc_prometheus.StreamServerInterceptor,
-		),
-	)
-
-	var hg hardwareGetter
-	dataModelVersion := env.Get("DATA_MODEL_VERSION")
-	switch dataModelVersion {
-	case "1":
-		tc, err := tinkClient.TinkHardwareClient()
-		if err != nil {
-			logger.Fatal(err, "Failed to create the tink client")
-		}
-		hg = hardwareGetterTinkerbell{tc}
-		// add health check for tink?
-	default:
-		cc, err := cacherClient.New(*facility)
-		if err != nil {
-			logger.Fatal(err, "Failed to create the cacher client")
-		}
-		hg = hardwareGetterCacher{cc}
-		go func() {
-			c := time.Tick(15 * time.Second)
-			for range c {
-				// Get All hardware as a proxy for a healthcheck
-				// TODO (patrickdevivo) until Cacher gets a proper healthcheck RPC
-				// a la https://github.com/grpc/grpc/blob/master/doc/health-checking.md
-				// this will have to do.
-				// Note that we don't do anything with the stream (we don't read from it)
-				var isCacherAvailableTemp bool
-				ctx, cancel := context.WithCancel(context.Background())
-				_, err := cc.All(ctx, &cacher.Empty{})
-				if err == nil {
-					isCacherAvailableTemp = true
-				}
+	c := make(chan os.Signal)
+	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
+	var ret error
+	ctx, cancel := context.WithCancel(context.TODO())
+	var once sync.Once
+	var wg sync.WaitGroup
+	runGoroutine(&ret, cancel, &once, &wg, func() error {
+		return httpserver.Serve(ctx, l, gitRev, time.Now())
+	})
+	runGoroutine(&ret, cancel, &once, &wg, func() error {
+		return grpcserver.Serve(ctx, l)
+	})
+	runGoroutine(&ret, cancel, &once, &wg, func() error {
+		select {
+		case sig, ok := <-c:
+			if ok {
+				l.With("signal", sig).Info("received stop signal, gracefully shutting down")
 				cancel()
-
-				isCacherAvailableMu.Lock()
-				isCacherAvailable = isCacherAvailableTemp
-				isCacherAvailableMu.Unlock()
-
-				if isCacherAvailableTemp {
-					metrics.CacherConnected.Set(1)
-					metrics.CacherHealthcheck.WithLabelValues("true").Inc()
-					logger.With("status", isCacherAvailableTemp).Debug("tick")
-				} else {
-					metrics.CacherConnected.Set(0)
-					metrics.CacherHealthcheck.WithLabelValues("false").Inc()
-					metrics.Errors.WithLabelValues("cacher", "healthcheck").Inc()
-					logger.With("status", isCacherAvailableTemp).Error(err)
-				}
 			}
-
-		}()
-	}
-
-	grpcServer := grpc.NewServer(serverOpts...)
-	hegelServer = &server{
-		log:            logger,
-		hardwareClient: hg,
-		subscriptions:  make(map[string]*subscription),
-	}
-
-	hegel.RegisterHegelServer(grpcServer, hegelServer)
-
-	lis, err := net.Listen("tcp", fmt.Sprintf(":%d", port))
-	if err != nil {
-		err = errors.Wrap(err, "failed to listen")
-		logger.Error(err)
-		panic(err)
-	}
-
-	// Register grpc prometheus server
-	grpc_prometheus.Register(grpcServer)
-	ServeHTTP()
-
-	metrics.State.Set(metrics.Ready)
-	//Serving GRPC
-	logger.Info("serving grpc")
-	err = grpcServer.Serve(lis)
-	if err != nil {
-		logger.Fatal(err, "Failed to serve  grpc")
+		case <-ctx.Done():
+		}
+		return nil
+	})
+	l.Info("waiting")
+	wg.Wait()
+	if ret != nil {
+		l.Fatal(ret)
 	}
 }
 
-func registerCustomEndpoints(mux *http.ServeMux) error {
-	customEndpoints = env.Get("CUSTOM_ENDPOINTS", `{"/metadata":".metadata"}`)
-	if mux == nil {
-		mux = http.DefaultServeMux
-	}
-
-	endpoints := make(map[string]string)
-	err := json.Unmarshal([]byte(customEndpoints), &endpoints)
-	if err != nil {
-		return errors.Wrap(err, "error in parsing custom endpoints")
-	}
-	for endpoint, filter := range endpoints {
-		mux.HandleFunc(endpoint, getMetadata(filter))
-	}
-
-	return nil
+func runGoroutine(ret *error, cancel context.CancelFunc, once *sync.Once, wg *sync.WaitGroup, fn func() error) {
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		err := fn()
+		if err != nil {
+			// we only care about the first error
+			once.Do(func() {
+				*ret = err
+				cancel()
+			})
+		}
+	}()
 }
